+++ skills/media-enrichment/src/media_enrichment/__init__.py
@@ -1,7 +1,7 @@
 """media-enrichment Skill package.
 
-Version: 0.1.0-dev22
+Version: 0.1.0-dev23
 """
 
-__version__ = "0.1.0-dev22"
+__version__ = "0.1.0-dev23"
 SKILL_NAME = "media-enrichment"